Move api dir
