+++ lib/routes/index.js
@@ -5,7 +5,7 @@ module.exports = function(crowi, app) {
 
   const routes = {
     Admin: require('./admin')(crowi, app, form),
-    API: require('./API')(crowi, app, form),
+    API: require('./api')(crowi, app, form),
     Login: require('./login')(crowi, app, form),
     Me: require('./me')(crowi, app, form),
   }
